Split Toolshed.Log into per-fun mods
