+++ lib/toolshed/log_attach.ex
@@ -1,11 +1,13 @@
-defmodule Toolshed.Log do
-  @moduledoc """
-  Utilities for attaching and detaching to the log
+defmodule Toolshed.LogAttach do
+  @moduledoc ""
 
-  These utilities configure Elixir's console backend to attach
-  to the current group leader. This makes it work over `ssh` sessions
-  and play well with the IEx prompt.
-  """
+  # Utilities for attaching and detaching to the log
+  #
+  # These utilities configure Elixir's console backend to attach
+  # to the current group leader. This makes it work over `ssh` sessions
+  # and play well with the IEx prompt.
+
+  @process_name Toolshed.Log
 
   @doc """
   Attach the current session to the Elixir logger
@@ -27,52 +29,19 @@ defmodule Toolshed.Log do
   """
   @spec log_attach(keyword()) :: {:error, any} | {:ok, :undefined | pid}
   def log_attach(options \\ []) do
-    case Process.get(__MODULE__) do
+    case Process.get(@process_name) do
       nil ->
         all_options = Keyword.put(options, :device, Process.group_leader())
         backend = {Logger.Backends.Console, all_options}
 
         {:ok, pid} = GenServer.start(Toolshed.Log.Watcher, {Process.group_leader(), backend})
 
-        Process.put(__MODULE__, {pid, backend})
+        Process.put(@process_name, {pid, backend})
 
         Logger.add_backend({Logger.Backends.Console, all_options})
 
       _other ->
         {:error, :detach_first}
-    end
-  end
-
-  @doc """
-  Detach the current session from the Elixir logger
-  """
-  @spec log_detach :: :ok | {:error, :not_attached | :not_found}
-  def log_detach() do
-    case Process.get(__MODULE__) do
-      nil ->
-        {:error, :not_attached}
-
-      {pid, backend} ->
-        Process.delete(__MODULE__)
-        GenServer.stop(pid)
-        Logger.remove_backend(backend)
-    end
-  end
-
-  defmodule Watcher do
-    @moduledoc false
-    use GenServer
-
-    @impl GenServer
-    def init({watch_pid, backend}) do
-      Process.monitor(watch_pid)
-      {:ok, backend}
-    end
-
-    @impl GenServer
-    def handle_info({:DOWN, _ref, :process, _pid, _reason}, backend) do
-      _ = Logger.remove_backend(backend)
-      {:stop, :normal, backend}
     end
   end
 end
